Correct criteria builder implementation (#193)

+++ jpa/odata-jpa-metadata/src/test/java/com/sap/olingo/jpa/metadata/core/edm/mapper/impl/TestTimeInstantLongConverter.java
@@ -41,7 +41,7 @@ public class TestTimeInstantLongConverter {
   @Test
   public void checkConvertToDatabaseColumnReturnConvertedOnNonNull() {
     final Instant time = Instant.parse(FIVE_DAYS_LATER);
-    final Long act = cut.convertToDatabaseColumn(time);
+    final Long act = (Long) cut.convertToDatabaseColumn(time);
     assertEquals(FIVE_DAYS, act);
   }
